projeto finalizado

+++ apidrf/core/views.py
@@ -21,41 +21,26 @@ from rest_framework import status
 
 
 @api_view(['GET'])
-# @authentication_classes([SessionAuthentication, BasicAuthentication])
-@permission_classes([IsAuthenticated])
-def example_view(request, format=None):
-    if request.method == 'POST':
-        dados_usuario = request.data
-        user = authenticate(
-            request, username=dados_usuario['usuario'], password=dados_usuario['password'])
-        if user is None:
-            print('teste')
-        else:
-            login(request, user)
-        # return Response(user)
-
-
-@api_view(['GET'])
 @permission_classes([IsAuthenticated])
 def index(request):
-    return JsonResponse({"teste": "teste"}, status=200, safe=False)
+    if request.method == 'GET':
+        return JsonResponse({"index": "index"}, status=200, safe=False)
+    return JsonResponse({"Negado": "Este método não é permitido"}, status=401, safe=False)
 
 
-@api_view(['GET', 'POST'])
+@api_view(['POST'])
 def create_usuario(request):
-
     dados_usuario = request.data
     if request.method == 'POST':
         serializer = UsuarioSerializer(data=dados_usuario, many=False)
-        print(serializer)
         if serializer.is_valid():
             serializer.save()
-        return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+        return JsonResponse({'usuario': serializer.data}, status=200, safe=False)
 
 
-@api_view(['GET', 'DELETE'])
-def delete_usuario(request, id):
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+# @api_view(['GET', 'DELETE'])
+# def delete_usuario(request, id):
+#     return JsonResponse({'testando': 'testando'}, status=200, safe=False)
 
 
 @api_view(['GET'])
@@ -63,58 +48,41 @@ def delete_usuario(request, id):
 def view_sala(request, id):
     dados_sala = request.data
     if request.method == 'GET':
-        print(id)
         return JsonResponse({'200': '200'}, status=200, safe=False) if Sala.objects.filter(id=id).filter(entrar=True).exists() else JsonResponse({'404': '404'}, status=404, safe=False)
-
-        # serializer = CheckSala(data=dados_sala, many=False)
-        # if serializer.is_valid():
-        #     serializer.save()
-        # return JsonResponse({'testando':'testando'}, status=200, safe=False)
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+    return JsonResponse({'Error': 'Método não permitido'}, status=401, safe=False)
 
 @api_view(['GET'])
 @permission_classes([IsAuthenticated])
 def get_salas(request):
-    # dados_sala = request.data
     if request.method == 'GET':
         salas = SalaSerializer(Sala.objects.all(), many=True)
         return JsonResponse({'salas': salas.data}, status=200, safe=False)
-
-        # serializer = CheckSala(data=dados_sala, many=False)
-        # if serializer.is_valid():
-        #     serializer.save()
-        # return JsonResponse({'testando':'testando'}, status=200, safe=False)
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+    return JsonResponse({'Error': 'Método não permitido'}, status=401, safe=False)
 
 @api_view(['POST', 'GET'])
-# @permission_classes([IsAuthenticated])
+@permission_classes([IsAuthenticated])
 def create_sala(request):
     dados_sala = request.data
-    print(Sala.objects.all())
     if request.method == 'POST':
-        print(Sala.objects.all())
         serializer = SalaSerializer(data=dados_sala, many=False)
         if serializer.is_valid():
-            print('here')
             serializer.save()
-        return JsonResponse({'testando': 'testando'}, status=200, safe=False)
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+        return JsonResponse({'sala': serializer.data}, status=200, safe=False)
+    return JsonResponse({'Error': 'Método não permitido'}, status=401, safe=False)
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def delete_sala(request, id):
-    sala = Sala.objects.filter(id=id)
-    sala.delete()
-    # dados_sala = request.data
-    # if request.method == 'POST':
-    #     serializer = SalaSerializer(data=dados_sala, many=False)
-    #     if serializer.is_valid():
-    #         serializer.save()
-    #     return JsonResponse({'testando':'testando'}, status=200, safe=False)
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+    if request.method == 'DELETE':
+        sala = Sala.objects.filter(id=id)
+        sala.delete()
+        return JsonResponse({'Success': 'Success'}, status=200, safe=False)
+    return JsonResponse({'Error': 'Método não permitido'}, status=401, safe=False)
 
 
 @api_view(['POST', 'GET'])
+@permission_classes([IsAuthenticated])
 def create_pergunta(request):
     dados_pergunta = request.data
     if request.method == 'POST':
@@ -122,89 +90,59 @@ def create_pergunta(request):
             titulo=dados_pergunta['titulo'], corpo=dados_pergunta['corpo'], tags=dados_pergunta['tags'])
         pergunta_sala = PerguntaSala.objects.create(
             id_pergunta=pergunta, id_sala=Sala.objects.get(id=dados_pergunta['id']))
-        print(PerguntaSala.objects.all())
-        # serializer = SalaSerializer(data=dados_sala, many=False)
-        # if serializer.is_valid():
-        #     serializer.save()
-        return JsonResponse({'testando': 'testando'}, status=200, safe=False)
-    if request.method == 'GET':
-        print(request.data)
-        # serializer = CheckSala(data=dados_sala, many=False)
-        # if serializer.is_valid():
-        #     serializer.save()
-        # return JsonResponse({'testando':'testando'}, status=200, safe=False)
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+        return JsonResponse({'Pergunta': 'Success'}, status=200, safe=False)
+    return JsonResponse({'Error': 'Método não permitido'}, status=401, safe=False)
 
 
-@api_view(['POST', 'GET', 'PUT'])
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
 def view_perguntas(request, id):
     if request.method == 'GET':
         perguntas = PerguntasSerializer(PerguntaSala.objects.select_related(
             'id_sala').filter(id_sala=id), many=True)
         return JsonResponse({'perguntas': perguntas.data}, status=200, safe=False)
-
+    return JsonResponse({'Error': 'Método não permitido'}, status=401, safe=False)
 
 @api_view(['PUT'])
+@permission_classes([IsAuthenticated])
 def edit_pergunta(request, id):
     if request.method == 'PUT':
         Perguntas.objects.filter(id=id).update(status=True)
-        print('entrei')
-    return JsonResponse({'perguntas': '201'}, status=201, safe=False)
+        return JsonResponse({'perguntas': '201'}, status=201, safe=False)
+    return JsonResponse({'Error': 'Método não permitido'}, status=401, safe=False)
 
 
 @api_view(['DELETE'])
+@permission_classes([IsAuthenticated])
 def delete_pergunta(request, id):
     if request.method == 'DELETE':
-        # print('entrei')
-        # pergunta = get_object_or_404(Perguntas, id=id)
-        # print(pergunta)
-        # pergunta.delete()
-        # print(Perguntas.objects.filter(id=id))
         Perguntas.objects.filter(id=id).delete()
-        # perguntas = PerguntasSerializer(Perguntas.objects.all(), many=True)
         return JsonResponse({'perguntas': 'deletada'}, status=200, safe=False)
     return JsonResponse({'perguntas': 'deletada'}, status=404, safe=False)
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated])
 def resposta_sala(request, id):
-    print(id)
     dados = request.data
     if request.method == 'POST':
-
         resposta = Resposta.objects.create(
             resposta=dados['resposta'], usuario=Usuario.objects.get(id=dados['usuario']))
         resposta_pergunta = PerguntaResposta.objects.create(
             id_pergunta=PerguntaSala.objects.get(id=id), id_resposta=resposta)
-        print(PerguntaResposta.objects.filter(id_pergunta=id))
-        return JsonResponse({'testando': 'testando'}, status=200, safe=False)
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
+        return JsonResponse({'resposta': 'resposta'}, status=200, safe=False)
+    return JsonResponse({'Error': 'error'}, status=401, safe=False)
 
 
 @api_view(['GET'])
+@permission_classes([IsAuthenticated])
 def get_respostas(request, id):
     dados = request.data
     if request.method == 'GET':
         respostas = RespostasSerializer(
             PerguntaResposta.objects.filter(id_pergunta=id), many=True)
         return JsonResponse({"respostas": respostas.data}, status=200, safe=False)
-    return JsonResponse({'testando': 'testando'}, status=200, safe=False)
-
-
-@api_view(['POST', 'GET'])
-def login_page(request):
-    if request.method == 'POST':
-        dados_usuario = request.data
-        user = authenticate(
-            request, username=dados_usuario['usuario'], password=dados_usuario['password'])
-        if user is None:
-            print('teste')
-        else:
-            login(request, user)
-            print(dados_usuario)
-
-    print(request.user)
-    return JsonResponse({'success': '201'}, status=201, safe=False)
+    return JsonResponse({'error': 'error'}, status=200, safe=False)
 
 
 @permission_classes([IsAuthenticated])
@@ -225,7 +163,3 @@ class LogoutView(generics.GenericAPIView):
             return Response(status=status.HTTP_205_RESET_CONTENT)
         except Exception as e:
             return Response(status=status.HTTP_400_BAD_REQUEST)
-
-
-def logout_page(request):
-    logout()
